+++ RELEASE_NOTES.md
@@ -14,8 +14,10 @@ This version adds PDF utilities and more print item types.
 
 * `Pdf` is a new namespace with PDF-specific types.
 * `PdfDataSource` is a new protocol that is implemented by `NSAttributedString`.
-* `PrintItem` has new `.imageData`, `.pdfData` and `.view` types.
-* `PrinterView` has a new `canPrintItem(:with:)` function.
+* `Printer` has new `canPrint` functions and properties.
+* `PrinterView` has new `canPrint` functions.
+* `PrintItem` has new `.imageData`, `.pdfData`, `.attributedString`, `.string` and `.view` types.
+
 
 ### 🚨 Breaking changes